requirements.txt
aiohttp
aio_pika
python-dotenv
beautifulsoup4
typing_extensions
